+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
 ## [Unreleased]
+### Changed
+- `step_based` config parameter deprecated in favor of `log_layout` parameter, by @nicke46
+### Added
+- RPHandler support for all Report Portal log levels, by @nicke46
+
+## [1.0.2]
 ### Added
 - `debug_mode` config parameter support, by @nicke46